+++ index.js
@@ -1,3 +1,5 @@
 export { default as eslintConfig } from "./eslint.config.js";
 export { default as prettierConfig } from "./prettier.config.cjs";
 export { default as stylelintConfig } from "./stylelint.config.cjs";
+export { default as typedocConfig } from "./typedoc.json" assert { type: "json" };
+export { default as typescriptConfig } from "./typescript.config.json" assert { type: "json" };